AP-23568: Enable Ask K-AI button if every non-optional connection is avail
Before, it would only check if the connection at index 1 is OK. However, this is
not valid for all nodes. The "Python Script" node can be configured to have no
input port and the "Variable Expression" node also functions if the input port
is not connected (because it is optional).

AP-23568 (scripting editor K-AI is not available if no inputs are connected, but for Python Script or Variable Expressions that can be fine)


(This commit was moved from https://github.com/knime/knime-scripting-editor. Original id: 98050418531e003a50aa082b31c55bfcdffa15d5)