UIEXT-2355: Prevent using Persist with Persistor or Migrate with Migration
In particular, it is no longer possible to use a custom persistor together with a config key provided by @Persist#configKey. This is by design, as it is more explicit and readable to define the config keys used by a custom persistor within that persistor (similar like a Migrator having to define what deprecated config keys it operates on, too).

UIEXT-2355 (Revisit Persistor API)