UIEXT-2375: Do not persist selected field in filters
While we do overwrite this field after it has been loaded before sending it as initial data to a view, we do not do that when sending it to the dialog. Since the dialog might trigger an inital view settings change on the view before the correct selected values are set by the respective jsonforms control, we rather handle undefined selected in the views.

UIEXT-2375 (Overwriting excluded manually selected in dialog Twinlist does not work)