UIEXT-2659: Adjust type displays
At least now the displayed types for flow vairable filters are the same as in the flow vairable port view. We might want to change both in the future to reference data type display names.

Also with this commit, the displayed types are sent to the frontend correctly again. Before that, the sent objects were empty since we cannot serialize records using JsonFormsDataUtil.

UIEXT-2659 (Adjust/streamline flow variable type displays)