UIEXT-1742: Enable optional widgets for most types
We do not yet support wrapping PersistableSettings (like e.g. the ColumnFilter) in an Optional, because it is not as easy to enable setting flow variables in this case.
In all other cases we now a) have to save a dummy value in case the optional is absent and b) send the current config key together with the ..._is_present config key as part of the persist schema. Both of these points are not easily generalized to arbitrary persistable settings.

UIEXT-1742 (Implement optional attribute (checkbox next to label) for all labeled widgets)