UIEXT-2353: Hello Marc. These are the files we want to make API
Generally, we (i.e. Robin and I) think it would be best to have individual packages for individual types of widgets.
E.g.:
- number
  integer  + double, all the validations
- text
  text, text area, rich text + all the validations
- choices
  -> single vs multiple?
  -> types (string, column, dataType, enum, variable)
  -> widget type (e.g. value switch vs dropdown vs radio buttons)
- credentials
- ???

Common annotations in one package (name = ?)
- @Widget
- @Layout (but all the further layout interfaces/annotation in a subpackage layout)
- @Effect (but all the further effect interfaces/annotation in a subpackage effect)
- @ValueProvider/@ValueReference (but all the further update interfaces/annotations in a subpackage update)
- ???

We decided against moving @Persistor, @Persist, NodeSettingsPersistor, etc. but it is referenced in the TypedStringFilter
--> Move it (only API) too but make the package(s) x-internal:=true

Also some persistors/migrations (e.g. LegacyNameFilterPersistor) relied on being in the same packag as the POJOs (to access fields that are not public as they are not relevant in the model (e.g. pattern)). We want to add getters and add @noreference here.

We should think about how to name things:

- Widget (or control / parameter / configuration  --> does not fit for things that do not control anything; element -> not clear to the user what this is). Also what do we mean by it (what is the TextMessage or the SimpleButtonWidget (not a control))?
- Predicate (naming overlap with java.functions)
- ???

Naming already clarified:
Layout, Effect, Migration, Parameters (formely Settings),  ???

Let's try to split implementation from API wherever it is not yet done

- DefaultNodeSettings static methods should be moved to something that stays in core-ui (not possible otherwise)
- E.g. serializer/deserializer of Credentials can stay in core-ui

UIEXT-2353 (Move to-be-published Widget API to isolated plugin)