NXT-3948: Do not create DataServices on deactivate call
instead, only deactivate the ones currently present.

Creating one only to immediately deactivate it is problematic because
the deactivation may come from a node or view disposal, in which case
references such as the PortObject are already invalid (i.e. `null` in
this particular bug).

NXT-3948 (Creating component with detached view fails)