UIEXT-2817: Substitute dots by <dot> for webui dialogs
This fixes two things:
* Until the issue https://github.com/eclipsesource/jsonforms/issues/2153 is addressed, jsonforms does not support dots as keys within the json
* Our own mechanism for flow variables also concatinated and split keys using "." making it impossible to differentiate between a.b -> c and a -> b -> c where "->" denotes a nesting inside an object

Both of these issues are now resolved backend-side by substituting dots with <dot>. This means that now config keys with "." are supported if and only if they are substituted with to "<dot>" whenever given to the frontend and it also means that acutal config keys using the prase <dot> are not supported anymore. We assume that this is not used anywhere.

UIEXT-2817 (Enable dots in config keys)