AP-24959: Enable using custom persistors on Credentials
With 7cba753 we introduced a mechanism that prevented serialization at any point in time after the dialog has been opened. I.e. when a flow variable was set, the value of that flow variable was not serialized to the frontend and in particular deserialization did not respect that value.
To mitigate that, we introduced a mechanism with 313dec1 that saves the name of the flow variable when it is set to the frontend-data so that on deserialization, we can serialize to that value which was necessary for state providers depending on that flow variable to work.
With ed6bf19 we changed the serialization so that also when setting a flow variable, the value is serialized. But we did not remove the custom deserialization mechanism.

This commit reverts the deserialization mechanism since it is not working anymore as soon as one can set flow variables that are not a single credentials flow variable.

AP-24959 (WebUI-Migration SSH Connector)