AP-25234: Move husky commit hooks to root pnpm project
With the current setup, nano-staged fails to locate the eslint config in `js-src/packages/scripting-editor`. Therefore, the pre-commit hook is configured only for @knime/core-ui. The configs for eslint, stylelint, and prettier should be consolidated to be able to apply the pre-commit hook to all packages.

AP-25234 (Allow usage of JSONForms dialogs generated from NodeParameters in Scripting Editors)