UIEXT-3127: Prevent tabs rendering before connectedFSOptions is provided
This fixes a bug where the file system tabs would render before the
"CONNECTED" file system was present in the list of available file
systems. Meaning that although the SideDrawerContent component would try
to display the "CONNECTED" file system when opened, it would not find it
and would default instead to the first available one (e.g. "LOCAL").

Now, if useFileChooserBrowseOptions detects that we indeed have a state
provided connectedFSOptions, then it uses a flag to only allow rendering
of the tab once the value has been provided.
Test changes:
* now that tabBar depends indirectly on a onMount call,
  SideDrawerContent tests must flush promises

UIEXT-3127 (WebUI-Migration Transfer Files)